Change in the part 'Getting your environment in sync'
When following the instructions in this section I got an error pushing my local changes to the Development environment. This was caused by the effect of step 8: removing the Document Types on the Development Environment before committing and pushing. 
When I pushed the changes before removing the Document Types I had to deploy the Development environment again from Kudu, but everything went well without errors.